nouvelle version LocalDate

+++ src/main/java/fr/eni/spring/Projet02/dal/UtilisateurDAOImpl.java
@@ -39,7 +39,7 @@ public class UtilisateurDAOImpl implements UtilisateurDAO {
             u.setMotDePasse(rs.getString("mot_de_passe"));
             u.setAdmin(rs.getBoolean("admin"));
             Adresse adresse = new Adresse();
-            adresse.setId(rs.getLong("id"));
+            adresse.setId(rs.getInt("id"));
             u.setAdresse(adresse);
             return u;
         }
